+++ src/Column.php
@@ -8,101 +8,37 @@ use InvalidArgumentException;
 
 class Column
 {
-    /**
-     * The table instance the table column is associated to.
-     *
-     * @property \Okipa\LaravelTable\Table $table
-     */
+    /** @property \Okipa\LaravelTable\Table $table */
     public $table;
-    /**
-     * The database table the table column is associated to.
-     *
-     * @property string $databaseDefaultTable
-     */
+    /** @property string $databaseDefaultTable */
     public $databaseDefaultTable;
-    /**
-     * The database table the table column is associated to for the searching action.
-     *
-     * @property string $databaseSearchedTable
-     */
+    /** @property string $databaseSearchedTable */
     public $databaseSearchedTable;
-    /**
-     * The database column the table column is associated to.
-     *
-     * @property string $databaseDefaultColumn
-     */
+    /** @property string $databaseDefaultColumn */
     public $databaseDefaultColumn;
-    /**
-     * The database column the table column is associated to for the searching action.
-     *
-     * @property string $databaseSearchedColumns
-     */
+    /** @property string $databaseSearchedColumns */
     public $databaseSearchedColumns;
-    /**
-     * The table column sortable status.
-     *
-     * @property bool $isSortable
-     */
+    /** @property bool $isSortable */
     public $isSortable;
-    /**
-     * The table column custom title.
-     *
-     * @property string $title
-     */
+    /** @property string $title */
     public $title;
-    /**
-     * The table column datetime format.
-     *
-     * @property string $dateTimeFormat
-     */
+    /** @property string $dateTimeFormat */
     public $dateTimeFormat;
-    /**
-     * The table column button classes.
-     *
-     * @property array $buttonClasses
-     */
+    /** @property array $buttonClasses */
     public $buttonClasses;
-    /**
-     * The table column string length limitation.
-     *
-     * @property int $stringLimit
-     */
+    /** @property int $stringLimit */
     public $stringLimit;
-    /**
-     * The table column url.
-     *
-     * @property string $url
-     */
+    /** @property string $url */
     public $url;
-    /**
-     * The table column custom value closure.
-     *
-     * @property Closure $valueClosure
-     */
+    /** @property Closure $valueClosure */
     public $valueClosure;
-    /**
-     * The table column custom html output closure.
-     *
-     * @property Closure $htmlClosure
-     */
+    /** @property Closure $htmlClosure */
     public $htmlClosure;
-    /**
-     * The table column value prepended icon.
-     *
-     * @property string $icon
-     */
+    /** @property string $icon */
     public $icon;
-    /**
-     * The icon display status when no value is found.
-     *
-     * @property bool $displayIconWhenNoValue
-     */
+    /** @property bool $displayIconWhenNoValue */
     public $displayIconWhenNoValue;
-    /**
-     * The table column classes.
-     *
-     * @property array $classes
-     */
+    /** @property array $classes */
     public $classes;
 
     /**