+++ views/bootstrap/table.blade.php
@@ -1,5 +1,5 @@
-<div {{ classTag('table-container', $table->containerClasses) }}>
-    <table {{ classTag('table', $table->tableClasses) }}>
+<div{{ classTag('table-container', $table->containerClasses) }}>
+    <table{{ htmlAttributes($table->identifier ? ['id' => $table->identifier] : null) }}{{ classTag('table', $table->tableClasses) }}>
         @include('laravel-table::' . $table->theadComponentPath)
         @include('laravel-table::' . $table->tbodyComponentPath)
         @include('laravel-table::' . $table->tfootComponentPath)