+++ views/bootstrap/tfoot.blade.php
@@ -1,7 +1,6 @@
 <tfoot>
-    <tr {{ classTag($table->trClasses) }}>
-        <td {{ classTag('bg-light', $table->tdClasses) }}
-            {{ htmlAttributes($table->columnsCount() > 1 ? ['colspan' => $table->columnsCount()] : null) }}>
+    <tr{{ classTag($table->trClasses) }}>
+        <td{{ classTag('bg-light', $table->tdClasses) }}{{ htmlAttributes($table->columnsCount() > 1 ? ['colspan' => $table->columnsCount()] : null) }}>
             <div class="d-flex justify-content-between flex-wrap py-2">
                 {{-- create button --}}
                 @if($table->isRouteDefined('create'))