+++ views/bootstrap/thead.blade.php
@@ -1,17 +1,16 @@
 <thead>
     {{-- rows number / search --}}
     @if($table->rowsNumberSelectionActivation || ! $table->searchableColumns->isEmpty())
-        <tr {{ classTag($table->trClasses) }}>
-            <td {{ classTag('bg-light', $table->tdClasses) }}
-                {{ htmlAttributes($table->columnsCount() > 1 ? ['colspan' => $table->columnsCount()] : null) }}>
+        <tr{{ classTag($table->trClasses) }}>
+            <td{{ classTag('bg-light', $table->tdClasses) }}{{ htmlAttributes($table->columnsCount() > 1 ? ['colspan' => $table->columnsCount()] : null) }}>
                 <div class="d-flex flex-wrap justify-content-between py-2">
                     {{-- rows number selection --}}
                     @if($table->rowsNumberSelectionActivation)
                         <div class="px-3 py-1 rows-number-selection">
                             <form role="form" method="GET" action="{{ $table->route('index') }}">
-                                <input type="hidden" name="search" value="{{ $table->request->search }}">
-                                <input type="hidden" name="sortBy" value="{{ $table->request->sortBy }}">
-                                <input type="hidden" name="sortDir" value="{{ $table->request->sortDir }}">
+                                <input type="hidden" name="{{ $table->searchField }}" value="{{ $table->request->get($table->searchField) }}">
+                                <input type="hidden" name="{{ $table->sortByField }}" value="{{ $table->request->get($table->sortByField) }}">
+                                <input type="hidden" name="{{ $table->sortDirField }}" value="{{ $table->request->get($table->sortDirField) }}">
                                 @foreach($table->appendedHiddenFields as $appendedKey => $appendedValue)
                                     <input type="hidden" name="{{ $appendedKey }}" value="{{ $appendedValue }}">
                                 @endforeach
@@ -23,8 +22,8 @@
                                     </div>
                                     <input class="form-control"
                                            type="number"
-                                           name="rows"
-                                           value="{{ $table->request->rows }}"
+                                           name="{{ $table->rowsField }}"
+                                           value="{{ $table->request->get($table->rowsField) }}"
                                            placeholder="@lang('laravel-table::laravel-table.rowsNumber')"
                                            aria-label="@lang('laravel-table::laravel-table.rowsNumber')">
                                     <div class="input-group-append">
@@ -42,9 +41,9 @@
                     @if(count($table->searchableColumns))
                         <div class="flex-fill px-3 py-1 searching">
                             <form role="form" method="GET" action="{{ $table->route('index') }}">
-                                <input type="hidden" name="rows" value="{{ $table->request->rows }}">
-                                <input type="hidden" name="sortBy" value="{{ $table->request->sortBy }}">
-                                <input type="hidden" name="sortDir" value="{{ $table->request->sortDir }}">
+                                <input type="hidden" name="{{ $table->rowsField }}" value="{{ $table->request->get($table->rowsField) }}">
+                                <input type="hidden" name="{{ $table->sortByField }}" value="{{ $table->request->get($table->sortByField) }}">
+                                <input type="hidden" name="{{ $table->sortDirField }}" value="{{ $table->request->get($table->sortDirField) }}">
                                 @foreach($table->appendedHiddenFields as $appendedKey => $appendedValue)
                                     <input type="hidden" name="{{ $appendedKey }}" value="{{ $appendedValue }}">
                                 @endforeach
@@ -57,17 +56,17 @@
                                     <input class="form-control"
                                            type="text"
                                            name="search"
-                                           value="{{ $table->request->search }}"
+                                           value="{{ $table->request->get($table->searchField) }}"
                                            placeholder="@lang('laravel-table::laravel-table.search') {{ $table->searchableTitles() }}"
                                            aria-label="@lang('laravel-table::laravel-table.search') {{ $table->searchableTitles() }}">
-                                    @if($table->request->search)
+                                    @if($table->request->get($table->searchField))
                                         <div class="input-group-append">
                                             <a class="input-group-text btn btn-link text-danger cancel-search"
                                                href="{{ $table->route('index', array_merge([
-                                                    'search'    => null,
-                                                    'rows'      => $table->request->rows,
-                                                    'sortBy'    => $table->request->sortBy,
-                                                    'sortDir'   => $table->request->sortDir
+                                                    $table->searchField    => null,
+                                                    $table->rowsField      => $table->request->get($table->rowsField),
+                                                    $table->sortByField    => $table->request->get($table->sortByField),
+                                                    $table->sortDirField   => $table->request->get($table->sortDirField)
                                                 ], $table->appendedValues)) }}"
                                                title="@lang('laravel-table::laravel-table.cancelSearch')">
                                                 <span>{!! config('laravel-table.icon.cancel') !!}</span>
@@ -91,23 +90,23 @@
         </tr>
     @endif
     {{-- column titles --}}
-    <tr {{ classTag($table->trClasses) }}>
+    <tr{{ classTag($table->trClasses) }}>
         @foreach($table->columns as $column)
-            <th {{ classTag($table->thClasses) }} scope="col">
+            <th{{ classTag($table->thClasses) }} scope="col">
                 @if($column->isSortable)
                     <a class="d-flex"
                        href="{{ $table->route('index', array_merge([
-                            'sortBy'    => $column->databaseDefaultColumn,
-                            'sortDir'   => $table->request->sortDir === 'desc' ? 'asc' : 'desc',
-                            'search'    => $table->request->search,
-                            'rows'      => $table->request->rows
+                            $table->rowsField      => $table->request->get($table->rowsField),
+                            $table->searchField    => $table->request->get($table->searchField),
+                            $table->sortByField    => $column->databaseDefaultColumn,
+                            $table->sortDirField   => $table->request->get($table->sortDirField) === 'desc' ? 'asc' : 'desc',
                         ], $table->appendedValues)) }}"
                        title="{{ $column->title }}">
-                        @if($table->request->sortBy === $column->databaseDefaultColumn 
-                            && $table->request->sortDir === 'asc')
+                        @if($table->request->get($table->sortByField) === $column->databaseDefaultColumn
+                            && $table->request->get($table->sortDirField) === 'asc')
                             <span class="sort asc">{!! config('laravel-table.icon.sortAsc') !!}</span>
-                        @elseif($table->request->sortBy === $column->databaseDefaultColumn 
-                            && $table->request->sortDir === 'desc')
+                        @elseif($table->request->get($table->sortByField) === $column->databaseDefaultColumn
+                            && $table->request->get($table->sortDirField) === 'desc')
                             <span class="sort desc">{!! config('laravel-table.icon.sortDesc') !!}</span>
                         @else
                             <span class="sort">{!! config('laravel-table.icon.sort') !!}</span>
@@ -122,7 +121,7 @@
             </th>
         @endforeach
         @if($table->isRouteDefined('edit') || $table->isRouteDefined('destroy'))
-            <th {{ classTag('text-right', $table->thClasses) }} scope="col">
+            <th{{ classTag('text-right', $table->thClasses) }} scope="col">
                 @lang('laravel-table::laravel-table.actions')
             </th>
         @endif